Upgrades people, upgrades
Updated tailwindcss to 3.1.8 from 3.0.6.

Rebuilt the javascript so now the website is semi-automatic. It scrapes apples event website for the latest event information regarding the time and date of the event and also updates the background to the new event artworks from apple.

All artwork is courtesy and property of  Apple Inc.